+++ electrumq/db/__init__.py
@@ -1,2 +1,8 @@
 # -*- coding: utf-8 -*-
 __author__ = 'zhouqi'
+
+"""
+
+sqlite下包括对block和tx的模型的存储
+
+"""
