--- electrumq/net/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-import logging
-
-__author__ = 'zhouqi'
-
-logger = logging.getLogger('network')
