+++ electrumq/network/ioloop.py
@@ -4,7 +4,7 @@ import time
 
 from tornado.ioloop import IOLoop as TornadoIOLoop, PeriodicCallback
 
-from electrumq.net import logger
+from electrumq.network import logger
 
 __author__ = 'zhouqi'
