+++ electrumq/secret/__init__.py
@@ -1,2 +1,13 @@
 # -*- coding: utf-8 -*-
 __author__ = 'zhouqi'
+
+"""
+
+私钥管理，重构中。应当包括：
+
+1. 私钥的保存与读取；
+2. 私钥的加密与解密；
+3. 私钥签名和公钥验证；
+4. hd等模式的支持（bip32）；
+
+"""
