+++ electrumq/tx/__init__.py
@@ -1,2 +1,13 @@
 # -*- coding: utf-8 -*-
 __author__ = 'zhouqi'
+
+"""
+
+tx=transaction，包括了：
+
+1. tx、input、output、script等模型
+2. coin chooser
+3. tx fee 预估模型
+
+
+"""
